+++ Source/EdenUniverseBuilder/Debug.h
@@ -16,4 +16,6 @@ public:
 	bool GetLogLevelFromString(FString Level, int& IntLevel);
 	bool SetLogLevel(FString Level);
 	bool Log(FString Message, FString Level);
+	bool LogFloat(FString MessageBefore, float Variable, FString MessageAfter, FString Level);
+	bool LogInt(FString MessageBefore, int Variable, FString MessageAfter, FString Level);
 };